Split out emitter logic into its own file, update examples, make alpha export minimal
