--- packages/dds/tree/src/domains/fence.json
@@ -1,5 +1,0 @@
-{
-	"tags": ["domains"],
-	"exports": ["index"],
-	"imports": ["util", "core", "feature-libraries"]
-}
