--- packages/dds/tree/src/events/fence.json
@@ -1,5 +1,0 @@
-{
-	"tags": ["events"],
-	"exports": ["index"],
-	"imports": ["util"]
-}
