+++ packages/dds/tree/src/events/index.ts
@@ -5,15 +5,16 @@
 
 export {
 	createEmitter,
+	createEmitterMinimal,
 	EventEmitter,
-	ComposableEventEmitter,
+	type IEmitter,
+	type NoListenersCallback,
+	type HasListeners,
+} from "./emitter.js";
+
+export {
 	type Listeners,
 	type Listenable,
 	type Off,
 	type IsListener,
-	type TransformListeners,
-	type UnionToIntersection,
-	type IEmitter,
-	type NoListenersCallback,
-	type HasListeners,
-} from "./events.js";
+} from "./listeners.js";
